[NestedStack] Move the compute fleet stack to its own module
